+++ onnxruntime/test/python/onnxruntime_test_ort_trainer.py
@@ -57,9 +57,10 @@ def generate_sample_batch(desc, batch_size, device):
     sample = generate_sample(desc_, device)
     return sample
 
-def runBertTrainingTest(gradient_accumulation_steps, use_mixed_precision, allreduce_post_accumulation):
+def runBertTrainingTest(gradient_accumulation_steps, use_mixed_precision, allreduce_post_accumulation,
+    use_simple_model_desc=True):
     model_desc = bert_model_description()
-    simple_model_desc = remove_extra_info(model_desc)
+    simple_model_desc = remove_extra_info(model_desc) if use_simple_model_desc else model_desc
     learning_rate_description = ort_trainer_learning_rate_description()
     device = torch.device("cuda", 0)
 
@@ -106,7 +107,7 @@ def runBertTrainingTest(gradient_accumulation_steps, use_mixed_precision, allred
 
         learning_rate = torch.tensor([lr]).to(device)
         if use_mixed_precision:
-            loss_scale = torch.tensor(loss_scaler.loss_scale_).to(device)
+            loss_scale = torch.tensor([loss_scaler.loss_scale_]).to(device)
             actual_loss = model.train_step(input_ids, segment_ids, input_mask, masked_lm_labels, next_sentence_labels, learning_rate, loss_scale)
             if isinstance(actual_loss, (list, tuple)):
                 assert len(actual_loss) == 2
