--- tools/ci_build/github/azure-pipelines/orttraining-linux-gpu-python-frontend-e2e-test-ci-pipeline.yml
@@ -1,35 +1,0 @@
-# 'trigger: none' to use UI to schedule the pipeline runs.
-# alternatively to schedule the pipeline run according to:
-# https://docs.microsoft.com/en-us/azure/devops/pipelines/process/scheduled-triggers?view=azure-devops&tabs=yaml
-trigger: none
-
-jobs:
-- job: OrtTraining_Linux_GPU_Python_FrontEnd_E2E
-
-  timeoutInMinutes: 120
-  
-  # alternatively we can configure (override) it directly from the build definition and not specify it here
-  pool: 'ORT-Training-Frontend-Linux-GPU'
-
-  steps:
-  - checkout: self
-    clean: true
-    submodules: recursive
-  
-  - template: templates/linux-set-variables-and-download.yml
-
-  # insert a python frontend test data preparation step here
-
-  - script: >
-      tools/ci_build/github/linux/run_dockerbuild.sh
-      -o ubuntu16.04 -d gpu -r $(Build.BinariesDirectory)
-      -x "
-      --enable_training
-      --config RelWithDebInfo
-      --skip_onnx_tests
-      --build_wheel
-      --enable_training_python_frontend_e2e_tests
-      "
-    displayName: 'Build and run frontend tests'
-
-  - template: templates/clean-agent-build-directory-step.yml
